net: Add new interface to tell usock available
Applications can prohibit Socket for INET by UsrSock by using the
ioctl command of SIOCDENYINETSOCK.

In this modification, we have added an interface to UsrSock that
queries whether or not the creation of Socket for INET is
prohibited.

This allows an application to choose which to use when both
UsrSock and, for example, rndis stacks are enabled.