cxd56xx/cxd56_audio: Fix an issue when driver is stopping
The audio driver release an audio buffer to user application via message
queue. A message queue can be used in irq context, but it has limitation
of depth of the queue.
So the driver stacked when the driver is stopping with many audio buffers
over the limitation, because it tries to release all audio buffers by
mq_send.

To avoid this behavior, change that releasing audio buffers are done in
LPWORKER.