sched/clock: Fix bug with system not booting by clock_systime_ticks()
This issue is caused by the following pull request.
https://github.com/apache/nuttx/pull/15139

This commit uses the RTC time to get the system tick, therefore the system tick is 0
until the RTC becomes available.

The cxd56xx system waits with the watchdog timer for the RTC to be available, but
if the tick is not incremented, RTC is permanently unavailable and the system cannot boot.

As a workaround, revert it and behave the same as before.